+++ core/model/event.go
@@ -81,7 +81,8 @@ type GroupEventFilter struct {
 	Limit  *int64 `json:"limit,omitempty"`
 	Offset *int64 `json:"offset,omitempty"`
 
-	StartTimeAfter *int64 `json:"start_time_after,omitempty"`
-
+	StartTimeAfter  *int64 `json:"start_time_after,omitempty"`
 	StartTimeBefore *int64 `json:"start_time_before,omitempty"`
+	EndTimeAfter    *int64 `json:"end_time_after,omitempty"`
+	EndTimeBefore   *int64 `json:"end_time_before,omitempty"`
 } // @name GroupEventFilter